:recycle: merge the two widgets
